+++ Main.py
@@ -1,0 +1,18 @@
+import sys
+from Acceso import *
+from tkinter import *
+from tkinter import ttk
+
+if __name__ == '__main__':
+    """ CREACIÓN DE LA VENTANA DE ACCESO AL SISTEMA"""
+
+    #Declara ventana de aplicación
+    root = Tk()
+
+    #LLamado a la aplicación mediante la ventana de acceso al sistema
+    #Access(root)
+
+    EraseSubject(root,1)# prueba de nueva ventana
+
+    #Bucle de la aplicación
+root.mainloop()
